UI: Add Okay/Cancel Buttons to Properties Dialog
Use QDialogButtonBox to add "Okay" and "Cancel" buttons to the
properties dialog.  The core functionality of the dialog is not changed;
I.E. the settings are still applied to the source as the user changes
them.  If the user clicks "Okay", the dialog simply exits.  If the user
clicks "Cancel", the original settings are reapplied to the source then
the dialog exits.  If the window is closed by any other means (I.E. by
the main obs window closing) then the properties dialog prompts the user
if they changed anything and asks if they wish to save their settings.

In order to implement this last feature, a method of checking for open
dialogs and sending each a quit message is added to the closeEvent()
method for OBSBasic.