UI: Add ability to change settings icons
Since it is impossible to access QListWidgetItems individually, using
the different types of Qt css selectors, I added the possibility.

The css used looks standard, but isn't. I parse the qss file upon theme
loading, and will initialize the needed icons. There is actually some
leeway in how the theme can format the icon sections, but most of it
breaks Qt's parser instead of ours.

The icons are created and set statically inside the theme class, since
there is no guarantee it will be loaded when we read a theme.