UI: Add SliderAbsoluteSetStyle class
This class, when used in conjuction with a QSlider control, allows
direct setting of the slider handle position, when clicking in an
area other than the slider handle.
The default QSlider handle behaviour is to step towards
clicked position.
This new QSlider behaviour is useful, for example, in the case
of volume slider where fast and 'precise' volume setting is
desired.
It can be also useful in other QSlider usage cases, for example,
in sources properties.