UI: Fix precision in float comparison helper
Use std::abs instead of the plain c abs function in order to avoid an
implicit cast to integer. This did cause the function to work as if the
epsilon parameter was rounded up to the next full integer value.
So with the default value of 0.01 it would in fact also return true for
all values with a difference of smaller than 1.